Move LocationDropdown into its own component
